# Use cached strings for ToString on integer values 0 to 9