@briansull 2018-06-08
The value 10 is also a commonly used, if you wan tto add one more to the cache.