@stephentoub 2018-06-08
> The value 10 is also a commonly used, if you wan to add one more to the cache.

We could experiment with that.  It would entail not only adding one more string but also changing the `bufferLength == 1` comparison to something more like `digits <= 1 && (uint)value <= 10`, or `bufferLength == 1 || (digits <= 2 && value == 10)`, or something like that.