@stephentoub 2018-06-09
I'm looking into the commonality of other values, and will follow-up if it looks fruitful.  For now, my goal was really just focused on 0 and 1, and 2-9 came along for the ride (cheaper to support them than to not support them).